# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "corpers.online API",
    "description": "Swagger documentation of our API endpoints.",
    "version": "1.0.0"
  },
  "host": "localhost:3051",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/v1.0/addplace": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/v1.0/signup": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "first_name": {
                  "example": "any"
                },
                "state_code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/api/v1.0/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1.0/chat": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "posts",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "502": {
            "description": "Bad Gateway"
          }
        }
      }
    },
...
  },
...
}